+++ ui/App.svelte
@@ -17,7 +17,7 @@
   import Project from "./Screen/Project.svelte";
   import Projects from "./Screen/Projects.svelte";
   import RegisterProject from "./Screen/RegisterProject.svelte";
-  import RegisterUser from "./Screen/RegisterUser.svelte";
+  import UserRegistration from "./Screen/UserRegistration.svelte";
   import Search from "./Screen/Search.svelte";
 
   initializeHotkeys();
@@ -47,7 +47,7 @@
     "/projects/:id/*": Project,
     "/design-system-guide": DesignSystemGuide,
     "/help": Help,
-    "/register-user": RegisterUser,
+    "/user-registration": UserRegistration,
     "*": NotFound
   };
 </script>
