+++ ui/Screen/UserRegistration.svelte
@@ -6,8 +6,8 @@
   import { Text, Title } from "../DesignSystem/Primitive";
   import { ModalLayout, StepCounter } from "../DesignSystem/Component";
 
-  import PickHandleStep from "./RegisterUser/PickHandleStep.svelte";
-  import SubmitRegistrationStep from "./RegisterUser/SubmitRegistrationStep.svelte";
+  import PickHandleStep from "./UserRegistration/PickHandleStep.svelte";
+  import SubmitRegistrationStep from "./UserRegistration/SubmitRegistrationStep.svelte";
 
   let step = 1;
   let handle = "cloudhead"; // TODO(merle): Get actual user profile (id, name, avatar)
